Move handle validation to form actions
