+++ src/components/Forms/FormActions.jsx
@@ -1,16 +1,50 @@
 import PropTypes from 'prop-types';
+import { PASSWORD_FORM, SINGLE_PAGE_FORM } from '../../constants/AppConstants';
+import { scrollToTop } from '../../utils/ScrollToElement';
+import Validator from '../../utils/Validator';
 
 const FormActions = ({
+  errorSummaryRef,
+  fields,
   formActions,
   formData,
-  handleValidation,
+  formType,
+  handleSubmit,
   isLoading,
   navigate,
+  setErrors,
 }) => {
   // Note: no forms currently require handleCancel
   const handleCancel = (redirectURL) => {
     sessionStorage.removeItem('formData');
     navigate(redirectURL);
+  };
+
+  const handleValidation = async (e, receivedFormData) => {
+    e.preventDefault();
+    const formErrors = await Validator({ formData: receivedFormData.formData, formFields: fields });
+    setErrors(formErrors);
+
+    if (formErrors.length < 1) {
+      /*
+       * Returning formData
+       * some forms perform special actions on the formData post validation
+       * e.g. CookiePolicy form will set cookie states
+       * so we always pass formData back
+       */
+      handleSubmit(receivedFormData);
+
+      /* If the form is a singlepage form we can clear the session
+       * we do not clear the session for multipage forms or sign in form
+       * as they have different needs
+      */
+      if (formType === SINGLE_PAGE_FORM || formType === PASSWORD_FORM) {
+        sessionStorage.removeItem('formData');
+      }
+    } else {
+      scrollToTop();
+      errorSummaryRef?.current?.focus();
+    }
   };
 
   return (
@@ -43,9 +77,13 @@ const FormActions = ({
 export default FormActions;
 
 FormActions.propTypes = {
+  errorSummaryRef: PropTypes.object.isRequired,
+  fields: PropTypes.array.isRequired,
   formActions: PropTypes.object.isRequired,
   formData: PropTypes.object.isRequired,
-  handleValidation: PropTypes.func.isRequired,
+  formType: PropTypes.string,
+  handleSubmit: PropTypes.func.isRequired,
   isLoading: PropTypes.bool,
   navigate: PropTypes.func.isRequired,
+  setErrors: PropTypes.func.isRequired,
 };
